Additional tests based on TestServer / TestClient
Created a TestFixture based on the TestServer, this can then be used as the
target for TestClient instances.  Currently a TestClient is spun up PER
test executed, which results in each test taking ~2s.
I've created a separate TestClientFixture which could be used as a Class
Fixture (https://xunit.net/docs/shared-context#class-fixture) to reduce the
time impact of TestClient instantiation, at the cost of less separation of
Client state between tests (within a Test Class).


Signed-off-by: Bevan Weiss <bevan.weiss@gmail.com>